+++ packages/frontend/src/components/chart/data-availability/da-absolute-throughput-chart.tsx
@@ -17,7 +17,7 @@ import { getXAxisProps } from '~/components/core/chart/get-x-axis-props'
 import { HorizontalSeparator } from '~/components/core/horizontal-separator'
 import { tooltipContentVariants } from '~/components/core/tooltip/tooltip'
 import type { DaThroughputDataPoint } from '~/server/features/data-availability/throughput/get-da-throughput-chart'
-import { formatBytes } from '~/server/features/data-availability/throughput/utils/formatBytes'
+import { formatBytes } from '~/server/features/data-availability/throughput/utils/format-bytes'
 import { formatTimestamp } from '~/utils/dates'
 
 interface Props {
