+++ packages/frontend/src/server/features/data-availability/throughput/utils/format-bytes.ts
@@ -1,6 +1,6 @@
 export function formatBytes(bytes: number) {
   if (bytes < 1024) {
-    return `${bytes} iB`
+    return `${bytes} B`
   }
   if (bytes < 1024 * 1024) {
     return `${(bytes / 1024).toFixed(2)} KiB`
